Display support for Advanced pause feature Added indicators and Coordinates to color themes Move to park point on manual pause Some visual improvements Better popup management Minor bug on load and unload menu items Powershell script to compile multiple configurations

+++ Compile.ps1
@@ -1,15 +1,15 @@
 cls
 Write-Output 'Clean Up' 
 Set-Location -Path 'D:\Users\Miguel\Documents\Trabajos\VSCode\Marlin_Ender3v2-Development'
-Rename-Item -Path '.\.pio\build' -NewName 'build-' -Force
-Remove-Item '.\.pio\build-' -Recurse
-C:\Users\Miguel\.platformio\penv\Scripts\platformio run --target clean -e STM32F103RET6_creality
+if (Test-Path -Path '.\.pio\build') { Rename-Item -Path '.\.pio\build' -NewName 'build-' -Force }
+if (Test-Path -Path '.\.pio\build-') { Remove-Item '.\.pio\build-' -Recurse }
+. $env:USERPROFILE\.platformio\penv\Scripts\platformio.exe run --target clean -e STM32F103RET6_creality
 Write-Output ' '
 Write-Output ' '
 #-------------------------------
 Write-Output 'Compiling 422 BLTouch ...'
 copy ".\configurations\v4.2.2\BLTouch\*.*" ".\Marlin"
-C:\Users\Miguel\.platformio\penv\Scripts\platformio run -e STM32F103RET6_creality
+. $env:USERPROFILE\.platformio\penv\Scripts\platformio.exe run -e STM32F103RET6_creality
 Write-Output 'Rename and move firmware'
 ls '.pio\build\STM32F103RET6_creality\*.bin' | Rename-Item -NewName {$_.name -replace "firmware","Experimental-422-BLTouch5x5"}
 Move-Item -Path '.pio\build\STM32F103RET6_creality\*.bin' -Destination '..\Releases'
@@ -18,7 +18,7 @@ Write-Output ' '
 #-------------------------------
 Write-Output 'Compiling 422 ManualMesh ...'
 copy ".\configurations\v4.2.2\ManualMesh\*.*" ".\Marlin"
-C:\Users\Miguel\.platformio\penv\Scripts\platformio run -e STM32F103RET6_creality
+. $env:USERPROFILE\.platformio\penv\Scripts\platformio.exe run -e STM32F103RET6_creality
 Write-Output 'Rename and move firmware'
 ls '.pio\build\STM32F103RET6_creality\*.bin' | Rename-Item -NewName {$_.name -replace "firmware","Experimental-422-ManualMesh5x5"}
 Move-Item -Path '.pio\build\STM32F103RET6_creality\*.bin' -Destination '..\Releases'
@@ -27,15 +27,15 @@ Write-Output ' '
 #-------------------------------
 Write-Output 'Clean Up'
 Start-Sleep -Seconds 2
-Rename-Item -Path '.\.pio\build' -NewName 'build-' -Force
-Remove-Item '.\.pio\build-' -Recurse
-C:\Users\Miguel\.platformio\penv\Scripts\platformio run --target clean -e STM32F103RET6_creality
+if (Test-Path -Path '.\.pio\build') { Rename-Item -Path '.\.pio\build' -NewName 'build-' -Force }
+if (Test-Path -Path '.\.pio\build-') { Remove-Item '.\.pio\build-' -Recurse }
+. $env:USERPROFILE\.platformio\penv\Scripts\platformio.exe run --target clean -e STM32F103RET6_creality
 Write-Output ' '
 Write-Output ' '
 #-------------------------------
 Write-Output 'Compiling 427 BLTouch ...'
 copy ".\configurations\v4.2.7\BLTouch\*.*" ".\Marlin"
-C:\Users\Miguel\.platformio\penv\Scripts\platformio run -e STM32F103RET6_creality
+. $env:USERPROFILE\.platformio\penv\Scripts\platformio.exe run -e STM32F103RET6_creality
 Write-Output 'Rename and move firmware'
 ls '.pio\build\STM32F103RET6_creality\*.bin' | Rename-Item -NewName {$_.name -replace "firmware","Experimental-427-BLTouch5x5"}
 Move-Item -Path '.pio\build\STM32F103RET6_creality\*.bin' -Destination '..\Releases'
@@ -44,7 +44,7 @@ Write-Output ' '
 #-------------------------------
 Write-Output 'Compiling 427 ManualMesh ...'
 copy ".\configurations\v4.2.7\ManualMesh\*.*" ".\Marlin"
-C:\Users\Miguel\.platformio\penv\Scripts\platformio run -e STM32F103RET6_creality
+. $env:USERPROFILE\.platformio\penv\Scripts\platformio.exe run -e STM32F103RET6_creality
 Write-Output 'Rename and move firmware'
 ls '.pio\build\STM32F103RET6_creality\*.bin' | Rename-Item -NewName {$_.name -replace "firmware","Experimental-427-ManualMesh5x5"}
 Move-Item -Path '.pio\build\STM32F103RET6_creality\*.bin' -Destination '..\Releases'
